Add JFOLStrat
